djchat/requirements.txt
Django==4.0.1
djangorestframework
psycopg2
channels
channels_redis
djangorestframework-simplejwt
djoser
django-cors-headers
Pillow